test(infinite-scroll): migrate top
